requirements.txt
streamlit
fastapi
uvicorn
pandas
numpy
plotly
statsmodels
prophet
pmdarima
pydantic
python-dotenv
reportlab
requests
supabase
openai
psycopg[binary]
ruff
pytest
